Update 7.0
BIG UPDATE: I added a simple main menu only the play and quit buttons work but there is a template options button. I added a second level that acts as the background of the main menu. I added functionality so that when 'E' is pressed the main menu level is opened and the main menu screen is presented.                                        Note* After pressing 'E' and the main menu pops up your mouse wont be on screen you need to press left shift and F1 to use your mouse.